Combine scripts, add englishDescription to questionItem
