+++ ines-website/.gitignore
@@ -24,8 +24,6 @@ dist-ssr
 *.sw?
 
 src/assets/statas
-# TODO: use labels_meta from app
-src/assets/labels_meta
 public/question_items
 public/surveys_data
 .vercel
